requirements.txt
requests
click
setuptools
mysql.connector